+++ packages/core/i18n.js
@@ -1,0 +1,2 @@
+module.exports = require('./dist/cjs/i18n/index.js');
+