+++ packages/core/tools.js
@@ -1,0 +1,2 @@
+module.exports = require('./dist/cjs/tools/index.js');
+