+++ packages/core/utils.js
@@ -1,0 +1,2 @@
+module.exports = require('./dist/cjs/utils/index.js');
+