+++ packages/cli/CHANGELOG.md
@@ -4,5 +4,6 @@
 
 ### Patch Changes
 
-- Updated dependencies [d8aaba6]
+- test bump
+- Updated dependencies
   - @sitecore-content-sdk/core@1.4.0