+++ packages/create-content-sdk-app/CHANGELOG.md
@@ -1,4 +1,4 @@
-# @sitecore-content-sdk/core
+# create-content-sdk-app
 
 ## 1.4.0