+++ packages/react/CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Patch Changes
 
-- Updated dependencies [d8aaba6]
+- test bump
+- Updated dependencies
+  - @sitecore-content-sdk/search@0.1.0
   - @sitecore-content-sdk/core@1.4.0
-  - @sitecore-content-sdk/search@1.4.0